Now has basic back propagation.
After the user trains the algorithm they are asked if they want to save it.
After they save it they can later draw it in the form of a graph for the user to interpret. 
The next thing to do is make back propagation that is actually useful however that will be more difficult and will require a large amount of research.