neural network
wrong name because of weird upload. 
There is now proper feed forwards in the way which I understand it. It uses  sqrt and sigmoid functions but may still be wrong. The back prop still needs to be done. 

The curve for the last curve is really drawn out so that it gives a value which will be in line with the number which is in the hundreds so it will give a number smaller than 1.