started other parts of the code
I've now started adding new things to the code to deal with the need for the neural network. While I've done little work recently on the actual neural network with the back propagation still needing to be done, I have made the feed forward algorithm work. 

The data feed is being done and will take data from a data generator. 
There is now a class for a data entry that will be in a list.