Data gen code
The random data generator now generates data that is compatible with the main code of the neural network. The main code will also now be able to read and correctly format the files from the random data generator. 
There is now a class for requests and the data is now entered into these classes for future use.