Edits for different inputs
I'm changing the code so it works for a large number of inputs and has fewer layer/all together is more malleable and more useful for the task. 

Each unit will have 32 CPU units and 32 RAM units. 

The network is being rewritten so that it can have the correct number of nodes and use them correctly.
Right now it doesn't work because of an error about setting the data from the Queue of data entries to the input nodes of the neural network.