10,10,19 changes to the run function
The neural network is now being made so that it is better for the many inputs and outputs for neural network. I've fixed errors that occur when the network is trained to do with it using sets of data that are inputted into the input nodes of the neural network.