Fixed
The code is fixed of all known errors. 
I now need to make it so the code can take and process the requests properly and come up with a good value. This will be the prelude to the reward and back prop system that will be all that is left for me to finish.