Chesterfield
I've named him Chesterfield and he is a good AI.  I've started to fix all the problems with the AI that were mainly in the forwards propagation part of the code. Soon he should be finished and then I'll take over the world *evil laugh*. Of course not Chesterfield is a good boy and is only designed to optimise 5G networks.