Chesterfield
This is the code which is makes up Chesterfield. It has to be used in conjunction with the resource simulator and the data needs to be generated. It is best to generate something around 100 requests to use it at maximum efficiency. There are still some error and there is a lot of commented out code. It isn't particularly smooth etc and the timer may not always the correct time value.